+++ noodles-bam/src/record/codec/encoder.rs
@@ -53,7 +53,7 @@ where
 
     // bin
     let alignment_end = record.alignment_end().transpose()?;
-    put_bin(dst, alignment_start, alignment_end)?;
+    put_bin(dst, alignment_start, alignment_end);
 
     // n_cigar_op
     let cigar = overflowing_put_cigar_op_count(dst, record)?;
@@ -128,22 +128,16 @@ where
     Ok(())
 }
 
-fn put_bin<B>(
-    dst: &mut B,
-    alignment_start: Option<Position>,
-    alignment_end: Option<Position>,
-) -> io::Result<()>
+fn put_bin<B>(dst: &mut B, alignment_start: Option<Position>, alignment_end: Option<Position>)
 where
     B: BufMut,
 {
     let bin = match (alignment_start, alignment_end) {
-        (Some(start), Some(end)) => region_to_bin(start, end)?,
+        (Some(start), Some(end)) => region_to_bin(start, end),
         _ => UNMAPPED_BIN,
     };
 
     dst.put_u16_le(bin);
-
-    Ok(())
 }
 
 fn overflowing_put_cigar_op_count<B, R>(dst: &mut B, record: &R) -> io::Result<Option<Cigar>>
@@ -181,7 +175,7 @@ where
 
 // § 5.3 "C source code for computing bin number and overlapping bins" (2021-06-03)
 #[allow(clippy::eq_op)]
-pub(crate) fn region_to_bin(alignment_start: Position, alignment_end: Position) -> io::Result<u16> {
+pub(crate) fn region_to_bin(alignment_start: Position, alignment_end: Position) -> u16 {
     let start = usize::from(alignment_start) - 1;
     let end = usize::from(alignment_end) - 1;
 
@@ -199,7 +193,8 @@ pub(crate) fn region_to_bin(alignment_start: Position, alignment_end: Position) 
         0
     };
 
-    u16::try_from(bin).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
+    // SAFETY: Truncate overflowing bin IDs.
+    bin as u16
 }
 
 #[cfg(test)]
@@ -395,14 +390,27 @@ mod tests {
     }
 
     #[test]
-    fn test_region_to_bin() -> Result<(), Box<dyn std::error::Error>> {
+    fn test_region_to_bin() -> Result<(), noodles_core::position::TryFromIntError> {
         let start = Position::try_from(8)?;
         let end = Position::try_from(13)?;
-        assert_eq!(region_to_bin(start, end)?, 4681);
+        assert_eq!(region_to_bin(start, end), 4681);
 
         let start = Position::try_from(63245986)?;
         let end = Position::try_from(63245986)?;
-        assert_eq!(region_to_bin(start, end)?, 8541);
+        assert_eq!(region_to_bin(start, end), 8541);
+
+        // https://github.com/zaeleus/noodles/issues/229
+        const DEPTH: usize = 5;
+        const MIN_SHIFT: usize = 14;
+        const MAX_POSITION: usize = 1 << (MIN_SHIFT + 3 * DEPTH);
+        const U16_SIZE: usize = 1 << 16;
+        const MAX_BIN_ID: usize = (1 << ((DEPTH + 1) * 3)) / 7;
+        const WINDOW_SIZE: usize = 1 << MIN_SHIFT;
+        let start =
+            Position::try_from((MAX_POSITION + 1) + (WINDOW_SIZE * (U16_SIZE - MAX_BIN_ID)))?;
+        dbg!(start);
+        let end = start;
+        assert_eq!(region_to_bin(start, end), 0);
 
         Ok(())
     }